deepfake_gui.py

"""
DeepFake Studio - ASCII Only Version
"""
import streamlit as st
from PIL import Image
import io
import time

# Set page configuration
st.set_page_config(
    page_title="DeepFake Studio",
    page_icon="DF",
    layout="wide",
    initial_sidebar_state="expanded"
)

# Main app
def main():
    # Sidebar
    with st.sidebar:
        st.title("DeepFake Studio")
        st.markdown("---")
        
        # Navigation
        st.subheader("Navigation")
        page = st.radio("", 
                     ["Face Swap", 
                      "Face Modification", 
                      "Face Morphing", 
                      "Emotion Transfer", 
                      "Diffusion Generation"])
        
        st.markdown("---")
        
        # About section
        with st.expander("About"):
            st.write("""
            DeepFake Studio is a powerful computer vision application 
            featuring state-of-the-art deepfake and face manipulation technologies.
            """)
    
    # Main content
    st.title("DeepFake Studio")
    st.markdown("Advanced facial manipulation powered by AI")
    
    # Face Swap
    if page == "Face Swap":
        st.header("Face Swap")
        st.markdown("Upload two images to swap the face from one to the other.")
        
        col1, col2 = st.columns(2)
        
        with col1:
            st.markdown("### Source Face")
            source_img = st.file_uploader("Upload source face", type=['jpg', 'jpeg', 'png'], key="source_face")
            if source_img:
                source_image = Image.open(source_img)
                st.image(source_image, caption="Source Image", use_column_width=True)
            
        with col2:
            st.markdown("### Target Image")
            target_img = st.file_uploader("Upload target image", type=['jpg', 'jpeg', 'png'], key="target_image")
            if target_img:
                target_image = Image.open(target_img)
                st.image(target_image, caption="Target Image", use_column_width=True)
        
        if source_img and target_img:
            if st.button("Perform Face Swap", key="face_swap_btn"):
                st.markdown("### Processing...")
                
                # Show a progress bar to simulate processing
                progress_bar = st.progress(0)
                status_text = st.empty()
                
                for i in range(101):
                    progress_bar.progress(i)
                    if i < 30:
                        status_text.text("Detecting faces...")
                    elif i < 60:
                        status_text.text("Aligning facial features...")
                    elif i < 90:
                        status_text.text("Applying face swap...")
                    else:
                        status_text.text("Finalizing results...")
                    time.sleep(0.02)
                
                status_text.text("Face swap completed! Implement your actual model here.")
                
                # For demo purposes - you'll replace this with your actual model output
                target_image = Image.open(target_img)
                st.image(target_image, caption="Face Swap Result (placeholder)", use_column_width=True)
                
                # Add a download button
                buf = io.BytesIO()
                target_image.save(buf, format="PNG")
                byte_im = buf.getvalue()
                
                st.download_button(
                    label="Download Result",
                    data=byte_im,
                    file_name="face_swap_result.png",
                    mime="image/png"
                )
    
    # Other features abbreviated for simplicity
    else:
        st.info(f"Selected {page} - Implement this feature by expanding the code.")
    
    # Footer
    st.markdown("---")
    st.markdown("DeepFake Studio 2025 | Built with Streamlit")

if __name__ == "__main__":
    main()
