requirements.txt
streamlit>=1.30.0
pillow>=9.0.0
numpy>=1.22.0
jinja2>=3.1.2